# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Vinylhound Backend API
  version: 1.0.0
  description: |
    The Vinylhound backend provides RESTful endpoints for authentication, catalog browsing,
    playlist management, favorites, and integrations with external music providers.
    Legacy `/api/*` routes are proxied to the same handlers but are not documented here.
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://api.vinylhound.example.com
    description: Production (replace with real hostname when available)
...
paths:
...
  /api/v1/auth/signup:
    post:
      tags:
        - Auth
      summary: Register a new user
      operationId: postSignup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignupRequest'
      responses:
        '201':
          description: Account created successfully
        '400':
          description: Invalid signup payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Username already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Unexpected server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/auth/login:
    post:
      tags:
        - Auth
      summary: Authenticate and retrieve a session token
      operationId: postLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authentication succeeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          description: Invalid login payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Unexpected server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/users/profile:
    get:
      tags:
        - Users
      summary: Retrieve the authenticated user's profile content
      operationId: getUserProfile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current content preferences
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserContent'
        '401':
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Unexpected server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      tags:
        - Users
      summary: Replace the authenticated user's profile content
      operationId: putUserProfile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserContent'
      responses:
        '204':
          description: Profile updated
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Unexpected server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/me/albums:
    get:
      tags:
        - Albums
      summary: List albums saved by the authenticated user
      operationId: getUserAlbums
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Albums owned by the user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlbumList'
        '401':
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Unexpected server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags:
        - Albums
      summary: Create a new album entry for the authenticated user
      operationId: postUserAlbum
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AlbumRequest'
      responses:
        '201':
          description: Album created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Album'
        '400':
          description: Invalid album payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Unexpected server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/me/albums/preferences:
    get:
      tags:
        - Albums
      summary: List album preferences for the authenticated user
      operationId: getUserAlbumPreferences
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Album preferences
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlbumPreferencesResponse'
        '401':
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Unexpected server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/v1/albums:
    get:
      tags:
        - Albums
      summary: Search albums in the global catalog
      operationId: getAlbums
      parameters:
        - name: artist
          in: query
          description: Filter by partial artist name
          schema:
            type: string
        - name: title
          in: query
          description: Filter by partial album title
          schema:
            type: string
        - name: genre
          in: query
          description: Filter by matching genre
          schema:
            type: string
        - name: year
          in: query
          description: Filter by release year
          schema:
            type: integer
        - name: rating
          in: query
          description: Filter by exact rating
          schema:
            type: integer
      responses:
        '200':
          description: Albums matching the filter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlbumList'
        '400':
          description: Invalid query parameter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Unexpected server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/v1/playlists:
    get:
      tags:
        - Playlists
      summary: List playlists owned by the authenticated user
      description: |
        Error responses currently return plain-text payloads.
      operationId: getPlaylists
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Playlists for the current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaylistList'
        '401':
          description: Missing or invalid bearer token
          content:
            text/plain:
              schema:
                type: string
                example: Authorization required
        '500':
          description: Unexpected server error
          content:
            text/plain:
              schema:
                type: string
    post:
      tags:
        - Playlists
      summary: Create a new playlist
      description: |
        Error responses currently return plain-text payloads.
      operationId: postPlaylist
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePlaylistRequest'
      responses:
        '201':
          description: Playlist created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Playlist'
        '400':
          description: Invalid playlist payload
          content:
            text/plain:
              schema:
                type: string
        '401':
          description: Missing or invalid bearer token
          content:
            text/plain:
              schema:
                type: string
                example: Authorization required
        '500':
          description: Unexpected server error
          content:
            text/plain:
              schema:
                type: string
...
  /api/v1/songs:
    get:
      tags:
        - Songs
      summary: Search songs
      description: |
        Error responses currently return plain-text payloads.
      operationId: getSongs
      parameters:
        - name: q
          in: query
          description: Free-text search across title and artist
          schema:
            type: string
        - name: artist
          in: query
          description: Filter by artist name
          schema:
            type: string
        - name: album
          in: query
          description: Filter by album title
          schema:
            type: string
        - name: album_id
          in: query
          description: Filter by album identifier
          schema:
            type: integer
      responses:
        '200':
          description: Songs matching the filter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SongList'
        '400':
          description: Invalid album_id
          content:
            text/plain:
              schema:
                type: string
                example: Invalid album_id
        '500':
          description: Unexpected server error
          content:
            text/plain:
              schema:
                type: string
...
  /api/v1/me/favorites/tracks:
    get:
      tags:
        - Favorites
      summary: List favorited track ids for the authenticated user
      operationId: getFavoriteTracks
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Favorited track references
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FavoriteTrackList'
        '401':
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Unexpected server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/v1/favorites/check:
    get:
      tags:
        - Favorites
      summary: (Reserved) Check if an item is favorited
      description: Placeholder endpoint that currently returns HTTP 501.
      operationId: getFavoritesCheck
      parameters:
        - name: song_id
          in: query
          schema:
            type: integer
            format: int64
          description: Song identifier to check
        - name: album_id
          in: query
          schema:
            type: integer
            format: int64
          description: Album identifier to check
      responses:
        '501':
          description: Not implemented
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
    SignupRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          minLength: 1
        password:
          type: string
          minLength: 1
        content:
          type: array
          items:
            type: string
          description: Optional list of preferred content tags
    LoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
        password:
          type: string
    TokenResponse:
      type: object
      required:
        - token
      properties:
        token:
          type: string
          description: JWT access token
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
    UserContent:
      type: object
      required:
        - content
      properties:
        content:
          type: array
          items:
            type: string
...
    AlbumRequest:
      type: object
      required:
        - artist
        - title
        - releaseYear
        - trackList
        - genreList
        - rating
      properties:
        artist:
          type: string
        title:
          type: string
        releaseYear:
          type: integer
        trackList:
          type: array
          items:
            type: string
        genreList:
          type: array
          items:
            type: string
        rating:
          type: integer
    AlbumList:
      type: object
      required:
        - albums
      properties:
        albums:
          type: array
          items:
            $ref: '#/components/schemas/Album'
...